int32 to int64

+++ api/src/main/kotlin/com/github/badoualy/telegram/api/utils/TLMessageUtils.kt
@@ -9,7 +9,7 @@ val TLAbsMessage?.date: Int
         else -> 0
     }
 
-val TLAbsMessage?.fromId: Int?
+val TLAbsMessage?.fromId: Long?
     get() = when (this) {
         is TLMessage -> fromId
         is TLMessageService -> fromId
